feat(alert): Alerting integration
Closes gravitee-io/issues#2777
