--- src/main/java/io/gravitee/repository/management/model/AlertCondition.java
@@ -1,77 +1,0 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.repository.management.model;
-
-import java.util.Objects;
-
-/**
- * @author Azize ELAMRANI (azize.elamrani at graviteesource.com)
- * @author GraviteeSource Team
- */
-public class AlertCondition {
-
-    private String metric;
-    private String threshold;
-    private int value;
-
-    public String getMetric() {
-        return metric;
-    }
-
-    public void setMetric(String metric) {
-        this.metric = metric;
-    }
-
-    public String getThreshold() {
-        return threshold;
-    }
-
-    public void setThreshold(String threshold) {
-        this.threshold = threshold;
-    }
-
-    public int getValue() {
-        return value;
-    }
-
-    public void setValue(int value) {
-        this.value = value;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof AlertCondition)) return false;
-        AlertCondition that = (AlertCondition) o;
-        return value == that.value &&
-                Objects.equals(metric, that.metric) &&
-                Objects.equals(threshold, that.threshold);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(metric, threshold, value);
-    }
-
-    @Override
-    public String toString() {
-        return "AlertCondition{" +
-                "metric='" + metric + '\'' +
-                ", threshold='" + threshold + '\'' +
-                ", value=" + value +
-                '}';
-    }
-}
